+++ apps/examples/desktop-app/webview/components/views/chat/chat-messages.tsx
@@ -1,39 +1,13 @@
 "use client";
 
-import { AgentApprovalCard, AgentAskQuestion } from "@cline/ui";
+import { AgentAskQuestion } from "@cline/ui";
 import {
-	Message as AgentMessage,
-	type AgentMessageRole,
 	Conversation,
 	ConversationContent,
 	ConversationScrollButton,
 	ConversationViewport,
-	MessageAction,
-	MessageActions,
-	MessageContent,
-	Reasoning,
-	ReasoningContent,
-	ReasoningTrigger,
-	ToolActivity,
-	ToolActivityCode,
-	ToolActivityContent,
-	ToolActivityDetails,
-	ToolActivityTrigger,
 } from "@cline/ui/components/agent-chat";
-import {
-	AlertCircle,
-	BrainIcon,
-	Check,
-	Clock3,
-	Copy,
-	Loader2,
-	PencilIcon,
-	ShieldAlert,
-	SplitIcon,
-	UndoIcon,
-	WrenchIcon,
-	X,
-} from "lucide-react";
+import { Clock3, Loader2 } from "lucide-react";
 import { memo, useCallback, useEffect, useMemo, useState } from "react";
 import {
 	AlertDialog,
@@ -45,7 +19,6 @@ import {
 	AlertDialogHeader,
 	AlertDialogTitle,
 } from "@/components/ui/alert-dialog";
-import { Button } from "@/components/ui/button";
 import { toast } from "@/hooks/use-toast";
 import type {
 	ChatMessage,
@@ -53,26 +26,21 @@ import type {
 	ChatSessionStatus,
 } from "@/lib/chat-schema";
 import { cn } from "@/lib/utils";
-import { MemoizedMarkdown } from "../../ui/markdown";
-import { formatChatMessageContent } from "./message-content";
-import {
-	EXPANDED_PANEL_RAIL_CLASS,
-	IS_DEBUG,
-	STREAMING_TITLE_CLASS,
-} from "./messages/constants";
+import { STREAMING_TITLE_CLASS } from "./messages/constants";
 import {
 	buildPreviousTimestampMap,
 	buildUserRunCountMap,
-	formatThoughtLabel,
 	getThoughtDurationMilliseconds,
 	groupChatMessages,
 } from "./messages/group-messages";
-import { getToolNameIcon } from "./messages/tool-icons";
+import { ChatImageLightbox } from "./messages/image-lightbox";
+import { MessageBubble } from "./messages/message-bubble";
 import {
-	buildGroupedToolLabel,
-	buildToolPresentation,
-	formatToolValue,
-} from "./messages/tool-summaries";
+	formatApprovalTimestamp,
+	ToolApprovalPanel,
+	type ToolApprovalRequestItem,
+} from "./messages/tool-approval-panel";
+import { ToolMessageBlock } from "./messages/tool-message-block";
 
 type ChatMessagesProps = {
 	sessionId: string | null;
@@ -101,18 +69,6 @@ type ChatMessagesProps = {
 		runCount: number,
 	) => void | Promise<void>;
 	onForkSession?: () => void | Promise<void>;
-};
-
-type ToolApprovalRequestItem = {
-	requestId: string;
-	sessionId: string;
-	createdAt: string;
-	toolCallId: string;
-	toolName: string;
-	input?: unknown;
-	iteration?: number;
-	agentId?: string;
-	conversationId?: string;
 };
 
 type AskQuestionRequestItem = {
@@ -687,37 +643,10 @@ function ChatMessagesImpl({
 			</ConversationViewport>
 			<ConversationScrollButton />
 			{visibleExpandedImage ? (
-				<div
-					aria-label="Expanded attachment"
-					aria-modal="true"
-					className="absolute inset-0 z-50 flex items-center justify-center bg-background/95 p-4 backdrop-blur-sm"
-					role="dialog"
-				>
-					<button
-						aria-label="Close expanded attachment"
-						className="absolute inset-0 cursor-zoom-out"
-						onClick={() => setExpandedImage(null)}
-						type="button"
-					/>
-					<div className="pointer-events-none relative z-10 flex h-full w-full items-center justify-center">
-						{/* biome-ignore lint/performance/noImgElement: User-provided data URLs cannot use Next's optimizer. */}
-						<img
-							alt="Expanded attachment"
-							className="max-h-full max-w-full rounded-lg object-contain shadow-2xl"
-							src={`data:${visibleExpandedImage.mediaType};base64,${visibleExpandedImage.data}`}
-						/>
-						<Button
-							aria-label="Close image viewer"
-							className="pointer-events-auto absolute right-0 top-0 rounded-full"
-							onClick={() => setExpandedImage(null)}
-							size="icon"
-							type="button"
-							variant="secondary"
-						>
-							<X className="h-4 w-4" />
-						</Button>
-					</div>
-				</div>
+				<ChatImageLightbox
+					image={visibleExpandedImage}
+					onClose={() => setExpandedImage(null)}
+				/>
 			) : null}
 			<AlertDialog
 				open={checkpointConfirmation !== null}
@@ -799,408 +728,6 @@ function ChatMessagesImpl({
 
 export const ChatMessages = memo(ChatMessagesImpl);
 
-function formatApprovalTimestamp(raw: string): string {
-	const parsed = new Date(raw);
-	if (Number.isNaN(parsed.getTime())) {
-		return "Pending now";
-	}
-	return parsed.toLocaleString();
-}
-
-function formatApprovalInput(input: unknown): string {
-	if (input == null) {
-		return "{}";
-	}
-	if (typeof input === "string") {
-		return input;
-	}
-	try {
-		return JSON.stringify(input, null, 2);
-	} catch {
-		return String(input);
-	}
-}
-
-function ToolApprovalPanel({
-	items,
-	pendingActions,
-	requestErrors,
-	onApprove,
-	onReject,
-}: {
-	items: ToolApprovalRequestItem[];
-	pendingActions: Record<string, "approving" | "rejecting">;
-	requestErrors: Record<string, string>;
-	onApprove: (requestId: string) => void;
-	onReject: (requestId: string) => void;
-}) {
-	return (
-		<section className="rounded-xl border border-amber-400/40 bg-amber-500/5 p-3">
-			<div className="flex items-center gap-2 text-sm font-medium text-foreground">
-				<ShieldAlert className="h-4 w-4 text-amber-500" />
-				Tool approval required
-			</div>
-			<p className="mt-1 text-xs text-muted-foreground">
-				Review each tool call and approve or reject it before execution.
-			</p>
-			<div className="mt-3 flex flex-col gap-2">
-				{items.map((item) => {
-					const pendingAction = pendingActions[item.requestId];
-					const error = requestErrors[item.requestId];
-					return (
-						<AgentApprovalCard
-							description={
-								<>
-									Request {item.requestId}
-									{item.iteration != null
-										? ` · Iteration ${item.iteration}`
-										: ""}
-								</>
-							}
-							detail={formatApprovalInput(item.input)}
-							error={error}
-							key={item.requestId}
-							meta={
-								<>
-									<Clock3 className="h-3 w-3" />
-									{formatApprovalTimestamp(item.createdAt)}
-								</>
-							}
-							onApprove={() => onApprove(item.requestId)}
-							onReject={() => onReject(item.requestId)}
-							responding={
-								pendingAction === "approving"
-									? "approve"
-									: pendingAction === "rejecting"
-										? "reject"
-										: undefined
-							}
-							title={item.toolName}
-						/>
-					);
-				})}
-			</div>
-		</section>
-	);
-}
-
-// Memoized with id-parameterized callbacks: during streaming only the message
-// object that received a delta changes identity, so all other bubbles skip
-// re-rendering (and re-running their Markdown pipeline) per flush.
-const MessageBubble = memo(function MessageBubble({
-	agentRole,
-	message,
-	runCount,
-	isStreaming = false,
-	onCopyMessage,
-	onExpandImage,
-	onEditMessage,
-	editDisabled = false,
-	editPending = false,
-	editError,
-	onRestoreCheckpoint,
-	restoreDisabled = false,
-	restorePending = false,
-	restoreError,
-	wasCopied = false,
-	onForkSession,
-	forkDisabled = false,
-	forkPending = false,
-	forkError,
-	isLastAssistantMessage = false,
-	reasoningContent,
-	reasoningRedacted,
-	thoughtDurationMilliseconds,
-}: {
-	agentRole: AgentMessageRole;
-	message: ChatMessage;
-	runCount?: number;
-	isStreaming?: boolean;
-	onCopyMessage?: (messageId: string, content: string) => void | Promise<void>;
-	onExpandImage?: (image: ChatMessageImage) => void;
-	onEditMessage?: (
-		messageId: string,
-		content: string,
-		runCount: number,
-	) => void | Promise<void>;
-	editDisabled?: boolean;
-	editPending?: boolean;
-	editError?: string;
-	onRestoreCheckpoint?: (
-		messageId: string,
-		runCount: number,
-	) => void | Promise<void>;
-	restoreDisabled?: boolean;
-	restorePending?: boolean;
-	restoreError?: string;
-	wasCopied?: boolean;
-	onForkSession?: (messageId: string) => void | Promise<void>;
-	forkDisabled?: boolean;
-	forkPending?: boolean;
-	forkError?: string;
-	isLastAssistantMessage?: boolean;
-	reasoningContent: string;
-	reasoningRedacted: boolean;
-	thoughtDurationMilliseconds?: number;
-}) {
-	const isUser = message.role === "user";
-	const isError = message.role === "error";
-	const checkpoint = message.meta?.checkpoint;
-	const displayContent = formatChatMessageContent(
-		message.role,
-		message.content,
-	);
-	const shouldRenderAssistantActions =
-		message.role === "assistant" &&
-		!isStreaming &&
-		!isError &&
-		Boolean(displayContent.trim()) &&
-		Boolean(onCopyMessage || onForkSession);
-	const shouldRenderUserActions =
-		isUser &&
-		Boolean(
-			onCopyMessage ||
-				checkpoint ||
-				(onEditMessage && runCount && displayContent.trim()),
-		);
-	const keepUserActionsVisible =
-		restorePending ||
-		editPending ||
-		Boolean(restoreError) ||
-		Boolean(editError);
-	const keepAssistantActionsVisible =
-		isLastAssistantMessage || forkPending || Boolean(forkError);
-
-	const messageDate = new Date(message.createdAt);
-	const hasValidMessageDate = !Number.isNaN(messageDate.getTime());
-	const messageTime = hasValidMessageDate
-		? messageDate.toLocaleTimeString(undefined, {
-				hour: "numeric",
-				minute: "2-digit",
-			})
-		: null;
-	const messageTimestamp = messageTime ? (
-		<time
-			className="shrink-0 whitespace-nowrap text-[11px] leading-none text-muted-foreground"
-			dateTime={messageDate.toISOString()}
-			title={messageDate.toLocaleString()}
-		>
-			{messageTime}
-		</time>
-	) : null;
-
-	// Spacing between blocks comes solely from the conversation list's `gap-2`
-	// and this content column's `gap-2`; blocks must not add their own margins.
-	return (
-		<AgentMessage className="relative flex flex-col gap-2" from={agentRole}>
-			<MessageContent className="flex min-w-0 flex-col gap-2 wrap-break-word">
-				{reasoningContent || reasoningRedacted ? (
-					<ReasoningBlock
-						content={reasoningContent}
-						durationMilliseconds={thoughtDurationMilliseconds}
-						redacted={reasoningRedacted}
-						streaming={isStreaming}
-					/>
-				) : null}
-
-				{message.images?.length ? (
-					<div className="grid max-w-2xl gap-2">
-						{message.images.map((image, index) => (
-							<button
-								aria-label={`Expand attachment ${index + 1}`}
-								className="cursor-zoom-in overflow-hidden rounded-lg border border-border bg-muted text-left transition-opacity hover:opacity-90 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-								key={image.id}
-								onClick={() => onExpandImage?.(image)}
-								type="button"
-							>
-								{/* biome-ignore lint/performance/noImgElement: User-provided data URLs do not have dimensions and cannot use Next's optimizer. */}
-								<img
-									alt={`Attachment ${index + 1}`}
-									className="max-h-56.25 max-w-56.25 object-contain"
-									src={`data:${image.mediaType};base64,${image.data}`}
-								/>
-							</button>
-						))}
-					</div>
-				) : null}
-
-				{displayContent ? (
-					<div className="min-w-0 max-w-full wrap-break-word">
-						<MemoizedMarkdown
-							content={displayContent}
-							streaming={isStreaming && message.role === "assistant"}
-						/>
-					</div>
-				) : null}
-			</MessageContent>
-
-			{shouldRenderUserActions ? (
-				<>
-					<MessageActions
-						className="absolute right-0 top-full z-10 -translate-y-1"
-						visible={keepUserActionsVisible}
-					>
-						{onCopyMessage ? (
-							<MessageAction
-								className="min-w-0 p-0"
-								label={wasCopied ? "Copied user message" : "Copy user message"}
-								onClick={() => void onCopyMessage(message.id, message.content)}
-								title={wasCopied ? "Copied" : "Copy message"}
-							>
-								{wasCopied ? (
-									<Check className="h-3.5 w-3.5" />
-								) : (
-									<Copy className="h-3.5 w-3.5" />
-								)}
-							</MessageAction>
-						) : null}
-						{onEditMessage && runCount && displayContent.trim() ? (
-							<MessageAction
-								className="min-w-0 p-0"
-								disabled={editDisabled || editPending}
-								label="Edit user message"
-								onClick={() =>
-									void onEditMessage(message.id, displayContent, runCount)
-								}
-								title="Edit message and restart from this point"
-							>
-								{editPending ? (
-									<Loader2 className="h-3.5 w-3.5 animate-spin" />
-								) : (
-									<PencilIcon className="h-3.5 w-3.5" />
-								)}
-							</MessageAction>
-						) : null}
-						{checkpoint ? (
-							<MessageAction
-								className="min-w-0 p-0"
-								disabled={restoreDisabled || restorePending}
-								label="Restore checkpoint"
-								onClick={() =>
-									void onRestoreCheckpoint?.(message.id, checkpoint.runCount)
-								}
-								title="Restore checkpoint"
-							>
-								{restorePending ? (
-									<Loader2 className="h-3.5 w-3.5 animate-spin" />
-								) : (
-									<UndoIcon className="h-3.5 w-3.5" />
-								)}
-							</MessageAction>
-						) : null}
-						{messageTimestamp}
-					</MessageActions>
-					{restoreError ? (
-						<div className="text-right text-xs text-destructive">
-							{restoreError}
-						</div>
-					) : null}
-					{editError ? (
-						<div className="text-right text-xs text-destructive">
-							{editError}
-						</div>
-					) : null}
-				</>
-			) : null}
-
-			{shouldRenderAssistantActions ? (
-				<MessageActions
-					className="absolute left-0 top-full z-10 -translate-y-1"
-					visible={keepAssistantActionsVisible}
-				>
-					{onCopyMessage ? (
-						<MessageAction
-							className="min-w-0 p-0"
-							label={
-								wasCopied
-									? "Copied assistant message"
-									: "Copy assistant message"
-							}
-							onClick={() => void onCopyMessage(message.id, message.content)}
-							title={wasCopied ? "Copied" : "Copy raw assistant output"}
-						>
-							{wasCopied ? (
-								<Check className="h-3 w-3" />
-							) : (
-								<Copy className="h-3 w-3" />
-							)}
-						</MessageAction>
-					) : null}
-					{onForkSession ? (
-						<MessageAction
-							className="min-w-0 p-0"
-							disabled={forkDisabled || forkPending}
-							label="Fork session"
-							onClick={() => void onForkSession(message.id)}
-							title="Fork session - copy full message history into a new session"
-						>
-							{forkPending ? (
-								<Loader2 className="h-3 w-3 animate-spin" />
-							) : (
-								<SplitIcon className="h-3 w-3 rotate-90" />
-							)}
-						</MessageAction>
-					) : null}
-					{messageTimestamp}
-					{forkError ? (
-						<span className="text-[11px] text-destructive">{forkError}</span>
-					) : null}
-				</MessageActions>
-			) : null}
-		</AgentMessage>
-	);
-});
-
-function ReasoningBlock({
-	content,
-	durationMilliseconds,
-	redacted,
-	streaming = false,
-}: {
-	content: string;
-	durationMilliseconds?: number;
-	redacted: boolean;
-	streaming?: boolean;
-}) {
-	const displayContent = content || (redacted ? "[redacted]" : "");
-	const label = streaming
-		? "Thinking"
-		: formatThoughtLabel(durationMilliseconds);
-	if (!displayContent) {
-		return null;
-	}
-
-	return (
-		<Reasoning className="my-0" isStreaming={streaming}>
-			<ReasoningTrigger
-				aria-label={label}
-				className="gap-2 py-1 text-sm text-muted-foreground transition-colors hover:text-foreground"
-			>
-				<BrainIcon aria-hidden="true" className="h-4 w-4 shrink-0" />
-				<span className={cn("font-medium", streaming && STREAMING_TITLE_CLASS)}>
-					{label}
-				</span>
-			</ReasoningTrigger>
-			<ReasoningContent
-				className={cn(
-					EXPANDED_PANEL_RAIL_CLASS,
-					// Prose reflows, so the X axis is pinned shut: `overflow-y-auto`
-					// alone would compute overflow-x to `auto` and let a long
-					// unbreakable token add a horizontal scrollbar.
-					"max-h-48 overflow-x-hidden overflow-y-auto",
-					"text-sm leading-relaxed text-muted-foreground",
-				)}
-			>
-				<MemoizedMarkdown
-					classNames="text-sm font-thin"
-					content={displayContent}
-					streaming={streaming}
-				/>
-			</ReasoningContent>
-		</Reasoning>
-	);
-}
-
 function pruneRequestMap<T extends string>(
 	prev: Record<string, T>,
 	activeRequestIds: Set<string>,
@@ -1216,112 +743,3 @@ function pruneRequestMap<T extends string>(
 	}
 	return hasRemoved ? next : prev;
 }
-
-// Memoized with element-wise comparison: the grouping pass wraps the same
-// message objects in fresh arrays every commit, so reference-comparing the
-// contents lets finished tool blocks skip re-rendering during streaming.
-const ToolMessageBlock = memo(
-	function ToolMessageBlock({ messages }: { messages: ChatMessage[] }) {
-		const presentations = messages.map(buildToolPresentation);
-		if (presentations.length === 0) return null;
-		const hasError = presentations.some(({ payload }) => payload?.isError);
-		const isRunning = presentations.some(({ inProgress }) => inProgress);
-		const label = buildGroupedToolLabel(presentations);
-		const icons = presentations.map(({ toolName }) =>
-			getToolNameIcon(toolName),
-		);
-		const firstIcon = icons[0] ?? WrenchIcon;
-		const Icon = icons.every((icon) => icon === firstIcon)
-			? firstIcon
-			: WrenchIcon;
-		const details = presentations.flatMap(({ message, summary }) =>
-			summary.details.map((detail) => ({
-				detail,
-				key: `${message.id}_${detail}`,
-			})),
-		);
-		const inputPreviews = IS_DEBUG
-			? presentations
-					.map(({ message, payload, toolName }) => ({
-						key: message.id,
-						toolName,
-						value: payload ? formatToolValue(payload.input) : "",
-					}))
-					.filter(({ value }) => Boolean(value))
-			: [];
-		const resultPreviews = presentations
-			.map(({ message, payload, toolName }) => ({
-				key: message.id,
-				toolName,
-				value: payload?.isError ? formatToolValue(payload.result) : "",
-			}))
-			.filter(({ value }) => Boolean(value));
-		const hasExpandedSections =
-			details.length > 0 ||
-			inputPreviews.length > 0 ||
-			resultPreviews.length > 0;
-		const diff = presentations.reduce(
-			(total, { summary }) => ({
-				additions: total.additions + (summary.diff?.additions ?? 0),
-				deletions: total.deletions + (summary.diff?.deletions ?? 0),
-			}),
-			{ additions: 0, deletions: 0 },
-		);
-
-		return (
-			<ToolActivity className="my-0" expandable={hasExpandedSections}>
-				<ToolActivityTrigger
-					additions={diff.additions || undefined}
-					deletions={diff.deletions || undefined}
-					icon={
-						hasError ? (
-							<AlertCircle className="size-4 text-destructive/80" />
-						) : (
-							<Icon className="size-4" />
-						)
-					}
-					label={
-						<span className={cn(isRunning && STREAMING_TITLE_CLASS)}>
-							{label}
-						</span>
-					}
-					showDisclosureIcon={false}
-					status={hasError ? "error" : isRunning ? "running" : "success"}
-				/>
-				<ToolActivityContent className={EXPANDED_PANEL_RAIL_CLASS}>
-					{details.length > 0 ? (
-						<ToolActivityDetails className="whitespace-pre-wrap">
-							{details.map(({ detail, key }) => (
-								<div key={key}>{detail}</div>
-							))}
-						</ToolActivityDetails>
-					) : null}
-					{inputPreviews.map((preview) => (
-						<div className="space-y-1" key={`input_${preview.key}`}>
-							<div className="text-[11px] uppercase tracking-wide text-muted-foreground/80">
-								{presentations.length > 1
-									? `${preview.toolName} input`
-									: "Input"}
-							</div>
-							<ToolActivityCode className="text-sm">
-								{preview.value}
-							</ToolActivityCode>
-						</div>
-					))}
-					{resultPreviews.map((preview) => (
-						<div
-							className="mt-1 break-words text-destructive"
-							key={`result_${preview.key}`}
-						>
-							{presentations.length > 1 ? `${preview.toolName}: ` : null}
-							{preview.value}
-						</div>
-					))}
-				</ToolActivityContent>
-			</ToolActivity>
-		);
-	},
-	(prev, next) =>
-		prev.messages.length === next.messages.length &&
-		prev.messages.every((message, index) => message === next.messages[index]),
-);
