+++ apps/examples/desktop-app/webview/components/views/chat/messages/message-bubble.tsx
@@ -1,0 +1,289 @@
+"use client";
+
+import {
+	Message as AgentMessage,
+	type AgentMessageRole,
+	MessageAction,
+	MessageActions,
+	MessageContent,
+} from "@cline/ui/components/agent-chat";
+import {
+	Check,
+	Copy,
+	Loader2,
+	PencilIcon,
+	SplitIcon,
+	UndoIcon,
+} from "lucide-react";
+import { memo } from "react";
+import type { ChatMessage, ChatMessageImage } from "@/lib/chat-schema";
+import { MemoizedMarkdown } from "../../../ui/markdown";
+import { formatChatMessageContent } from "../message-content";
+import { ReasoningBlock } from "./reasoning-block";
+
+// Memoized with id-parameterized callbacks: during streaming only the message
+// object that received a delta changes identity, so all other bubbles skip
+// re-rendering (and re-running their Markdown pipeline) per flush.
+export const MessageBubble = memo(function MessageBubble({
+	agentRole,
+	message,
+	runCount,
+	isStreaming = false,
+	onCopyMessage,
+	onExpandImage,
+	onEditMessage,
+	editDisabled = false,
+	editPending = false,
+	editError,
+	onRestoreCheckpoint,
+	restoreDisabled = false,
+	restorePending = false,
+	restoreError,
+	wasCopied = false,
+	onForkSession,
+	forkDisabled = false,
+	forkPending = false,
+	forkError,
+	isLastAssistantMessage = false,
+	reasoningContent,
+	reasoningRedacted,
+	thoughtDurationMilliseconds,
+}: {
+	agentRole: AgentMessageRole;
+	message: ChatMessage;
+	runCount?: number;
+	isStreaming?: boolean;
+	onCopyMessage?: (messageId: string, content: string) => void | Promise<void>;
+	onExpandImage?: (image: ChatMessageImage) => void;
+	onEditMessage?: (
+		messageId: string,
+		content: string,
+		runCount: number,
+	) => void | Promise<void>;
+	editDisabled?: boolean;
+	editPending?: boolean;
+	editError?: string;
+	onRestoreCheckpoint?: (
+		messageId: string,
+		runCount: number,
+	) => void | Promise<void>;
+	restoreDisabled?: boolean;
+	restorePending?: boolean;
+	restoreError?: string;
+	wasCopied?: boolean;
+	onForkSession?: (messageId: string) => void | Promise<void>;
+	forkDisabled?: boolean;
+	forkPending?: boolean;
+	forkError?: string;
+	isLastAssistantMessage?: boolean;
+	reasoningContent: string;
+	reasoningRedacted: boolean;
+	thoughtDurationMilliseconds?: number;
+}) {
+	const isUser = message.role === "user";
+	const isError = message.role === "error";
+	const checkpoint = message.meta?.checkpoint;
+	const displayContent = formatChatMessageContent(
+		message.role,
+		message.content,
+	);
+	const shouldRenderAssistantActions =
+		message.role === "assistant" &&
+		!isStreaming &&
+		!isError &&
+		Boolean(displayContent.trim()) &&
+		Boolean(onCopyMessage || onForkSession);
+	const shouldRenderUserActions =
+		isUser &&
+		Boolean(
+			onCopyMessage ||
+				checkpoint ||
+				(onEditMessage && runCount && displayContent.trim()),
+		);
+	const keepUserActionsVisible =
+		restorePending ||
+		editPending ||
+		Boolean(restoreError) ||
+		Boolean(editError);
+	const keepAssistantActionsVisible =
+		isLastAssistantMessage || forkPending || Boolean(forkError);
+
+	const messageDate = new Date(message.createdAt);
+	const hasValidMessageDate = !Number.isNaN(messageDate.getTime());
+	const messageTime = hasValidMessageDate
+		? messageDate.toLocaleTimeString(undefined, {
+				hour: "numeric",
+				minute: "2-digit",
+			})
+		: null;
+	const messageTimestamp = messageTime ? (
+		<time
+			className="shrink-0 whitespace-nowrap text-[11px] leading-none text-muted-foreground"
+			dateTime={messageDate.toISOString()}
+			title={messageDate.toLocaleString()}
+		>
+			{messageTime}
+		</time>
+	) : null;
+
+	// Spacing between blocks comes solely from the conversation list's `gap-2`
+	// and this content column's `gap-2`; blocks must not add their own margins.
+	return (
+		<AgentMessage className="relative flex flex-col gap-2" from={agentRole}>
+			<MessageContent className="flex min-w-0 flex-col gap-2 wrap-break-word">
+				{reasoningContent || reasoningRedacted ? (
+					<ReasoningBlock
+						content={reasoningContent}
+						durationMilliseconds={thoughtDurationMilliseconds}
+						redacted={reasoningRedacted}
+						streaming={isStreaming}
+					/>
+				) : null}
+
+				{message.images?.length ? (
+					<div className="grid max-w-2xl gap-2">
+						{message.images.map((image, index) => (
+							<button
+								aria-label={`Expand attachment ${index + 1}`}
+								className="cursor-zoom-in overflow-hidden rounded-lg border border-border bg-muted text-left transition-opacity hover:opacity-90 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
+								key={image.id}
+								onClick={() => onExpandImage?.(image)}
+								type="button"
+							>
+								{/* biome-ignore lint/performance/noImgElement: User-provided data URLs do not have dimensions and cannot use Next's optimizer. */}
+								<img
+									alt={`Attachment ${index + 1}`}
+									className="max-h-56.25 max-w-56.25 object-contain"
+									src={`data:${image.mediaType};base64,${image.data}`}
+								/>
+							</button>
+						))}
+					</div>
+				) : null}
+
+				{displayContent ? (
+					<div className="min-w-0 max-w-full wrap-break-word">
+						<MemoizedMarkdown
+							content={displayContent}
+							streaming={isStreaming && message.role === "assistant"}
+						/>
+					</div>
+				) : null}
+			</MessageContent>
+
+			{shouldRenderUserActions ? (
+				<>
+					<MessageActions
+						className="absolute right-0 top-full z-10 -translate-y-1"
+						visible={keepUserActionsVisible}
+					>
+						{onCopyMessage ? (
+							<MessageAction
+								className="min-w-0 p-0"
+								label={wasCopied ? "Copied user message" : "Copy user message"}
+								onClick={() => void onCopyMessage(message.id, message.content)}
+								title={wasCopied ? "Copied" : "Copy message"}
+							>
+								{wasCopied ? (
+									<Check className="h-3.5 w-3.5" />
+								) : (
+									<Copy className="h-3.5 w-3.5" />
+								)}
+							</MessageAction>
+						) : null}
+						{onEditMessage && runCount && displayContent.trim() ? (
+							<MessageAction
+								className="min-w-0 p-0"
+								disabled={editDisabled || editPending}
+								label="Edit user message"
+								onClick={() =>
+									void onEditMessage(message.id, displayContent, runCount)
+								}
+								title="Edit message and restart from this point"
+							>
+								{editPending ? (
+									<Loader2 className="h-3.5 w-3.5 animate-spin" />
+								) : (
+									<PencilIcon className="h-3.5 w-3.5" />
+								)}
+							</MessageAction>
+						) : null}
+						{checkpoint ? (
+							<MessageAction
+								className="min-w-0 p-0"
+								disabled={restoreDisabled || restorePending}
+								label="Restore checkpoint"
+								onClick={() =>
+									void onRestoreCheckpoint?.(message.id, checkpoint.runCount)
+								}
+								title="Restore checkpoint"
+							>
+								{restorePending ? (
+									<Loader2 className="h-3.5 w-3.5 animate-spin" />
+								) : (
+									<UndoIcon className="h-3.5 w-3.5" />
+								)}
+							</MessageAction>
+						) : null}
+						{messageTimestamp}
+					</MessageActions>
+					{restoreError ? (
+						<div className="text-right text-xs text-destructive">
+							{restoreError}
+						</div>
+					) : null}
+					{editError ? (
+						<div className="text-right text-xs text-destructive">
+							{editError}
+						</div>
+					) : null}
+				</>
+			) : null}
+
+			{shouldRenderAssistantActions ? (
+				<MessageActions
+					className="absolute left-0 top-full z-10 -translate-y-1"
+					visible={keepAssistantActionsVisible}
+				>
+					{onCopyMessage ? (
+						<MessageAction
+							className="min-w-0 p-0"
+							label={
+								wasCopied
+									? "Copied assistant message"
+									: "Copy assistant message"
+							}
+							onClick={() => void onCopyMessage(message.id, message.content)}
+							title={wasCopied ? "Copied" : "Copy raw assistant output"}
+						>
+							{wasCopied ? (
+								<Check className="h-3 w-3" />
+							) : (
+								<Copy className="h-3 w-3" />
+							)}
+						</MessageAction>
+					) : null}
+					{onForkSession ? (
+						<MessageAction
+							className="min-w-0 p-0"
+							disabled={forkDisabled || forkPending}
+							label="Fork session"
+							onClick={() => void onForkSession(message.id)}
+							title="Fork session - copy full message history into a new session"
+						>
+							{forkPending ? (
+								<Loader2 className="h-3 w-3 animate-spin" />
+							) : (
+								<SplitIcon className="h-3 w-3 rotate-90" />
+							)}
+						</MessageAction>
+					) : null}
+					{messageTimestamp}
+					{forkError ? (
+						<span className="text-[11px] text-destructive">{forkError}</span>
+					) : null}
+				</MessageActions>
+			) : null}
+		</AgentMessage>
+	);
+});
